refactor: move handle write to profile ui
